+++ src/ATV/index.ts
@@ -7,44 +7,7 @@ import { Command, CreateAndSendDocumentResponse } from './useCases/sendDocument/
 import * as parser from 'fast-xml-parser'
 import qs from 'querystring'
 import { ConfirmationMessageRaw } from '@src/types/facturaInterfaces'
-
-export type Mode = 'prod' | 'stg' | 'custom';
-
-export type ATVOptions = {
-  defaultOverwrites?: {
-    tokenServiceUrl: string;
-  };
-}
-
-export type TokenServiceProps = {
-  serviceUrl: string;
-  clientId: string;
-}
-
-export type SendResponse = {
-  status: number;
-  location: string;
-  errorCause?: string | null;
-}
-
-type SendConfirmationInput = {
-  url: string;
-  headers: {
-    Authorization: string;
-    'Content-Type': string;
-  };
-}
-
-type ConfirmationMessage = {
-  clave: string;
-  message: string;
-  details: string;
-  emitterFullName: string;
-  emitterIdentifierId: string;
-  receiverIdentifierId: string;
-  taxesTotalAmount: string;
-  totalInvoiceAmount: string;
-}
+import { ATVOptions, ConfirmationMessage, Mode, SendConfirmationInput, SendResponse } from './types'
 
 export class ATV {
   public readonly options: ATVOptions
